+++ src/signalling/room/command_handler.rs
@@ -60,7 +60,7 @@ impl CommandHandler for Room {
         })?;
 
         let event = if from_peer.is_known_to_remote() {
-            Event::TracksApplied {
+            Event::PeerUpdated {
                 peer_id: to_peer.id(),
                 negotiation_role: Some(NegotiationRole::Answerer(
                     sdp_offer.clone(),
@@ -195,7 +195,7 @@ impl CommandHandler for Room {
         Ok(())
     }
 
-    /// Sends [`Event::TracksApplied`] with data from the received
+    /// Sends [`Event::PeerUpdated`] with data from the received
     /// [`Command::UpdateTracks`].
     ///
     /// Starts renegotiation process.