+++ src/signalling/room/mod.rs
@@ -290,7 +290,7 @@ impl Room {
         }
     }
 
-    /// Sends [`Event::TracksApplied`] with latest [`Peer`] changes to specified
+    /// Sends [`Event::PeerUpdated`] with latest [`Peer`] changes to specified
     /// [`Member`]. Starts renegotiation, marking provided [`Peer`] as
     /// [`NegotiationRole::Offerer`].
     ///
@@ -313,7 +313,7 @@ impl Room {
 
         self.members.send_event_to_member(
             member_id,
-            Event::TracksApplied {
+            Event::PeerUpdated {
                 updates,
                 negotiation_role: Some(NegotiationRole::Offerer),
                 peer_id,