+++ src/signalling/room/peer_events_handler.rs
@@ -4,7 +4,7 @@ use actix::{Handler, Message, StreamHandler, WeakAddr};
 use chrono::{DateTime, Utc};
 use medea_client_api_proto::{
     ConnectionQualityScore, Event, MemberId, NegotiationRole, PeerId,
-    TrackUpdate,
+    PeerUpdate,
 };
 
 use crate::{
@@ -197,12 +197,12 @@ impl PeerUpdatesSubscriber for WeakAddr<Room> {
     }
 
     /// Upgrades [`WeakAddr`] and if it's successful then notifies [`Room`] that
-    /// the provided [`TrackUpdate`]s were forcibly (without negotiation)
+    /// the provided [`PeerUpdate`]s were forcibly (without negotiation)
     /// applied to the [`Peer`].
     ///
     /// If [`WeakAddr`] upgrade fails then nothing will be done.
     #[inline]
-    fn force_update(&self, peer_id: PeerId, changes: Vec<TrackUpdate>) {
+    fn force_update(&self, peer_id: PeerId, changes: Vec<PeerUpdate>) {
         if let Some(addr) = self.upgrade() {
             addr.do_send(ForceUpdate(peer_id, changes));
         }
@@ -210,18 +210,18 @@ impl PeerUpdatesSubscriber for WeakAddr<Room> {
 }
 
 /// [`Message`] which indicates that [`Peer`] with a provided [`PeerId`] should
-/// be updated with provided [`TrackUpdate`]s without negotiation.
+/// be updated with provided [`PeerUpdate`]s without negotiation.
 ///
 /// Can be done in any [`Peer`] state.
 #[derive(Message, Clone, Debug)]
 #[rtype(result = "Result<(), RoomError>")]
-pub struct ForceUpdate(PeerId, Vec<TrackUpdate>);
+pub struct ForceUpdate(PeerId, Vec<PeerUpdate>);
 
 impl Handler<ForceUpdate> for Room {
     type Result = Result<(), RoomError>;
 
     /// Gets [`MemberId`] of the provided [`Peer`] and sends all provided
-    /// [`TrackUpdate`]s to this [`MemberId`] with `negotiation_role: None`.
+    /// [`PeerUpdate`]s to this [`MemberId`] with `negotiation_role: None`.
     fn handle(
         &mut self,
         msg: ForceUpdate,
@@ -232,7 +232,7 @@ impl Handler<ForceUpdate> for Room {
             .map_peer_by_id(msg.0, PeerStateMachine::member_id)?;
         self.members.send_event_to_member(
             member_id,
-            Event::TracksApplied {
+            Event::PeerUpdated {
                 peer_id: msg.0,
                 updates: msg.1,
                 negotiation_role: None,
@@ -258,14 +258,14 @@ impl Handler<NegotiationNeeded> for Room {
     /// Sends [`Event::PeerCreated`] if this [`Peer`] unknown for the remote
     /// side.
     ///
-    /// Sends [`Event::TracksApplied`] if this [`Peer`] known for the remote
+    /// Sends [`Event::PeerUpdated`] if this [`Peer`] known for the remote
     /// side.
     ///
     /// If this [`Peer`] or it's partner not [`Stable`] then forcible
     /// track changes will be committed.
     ///
     /// [`Event::PeerCreated`]: medea_client_api_proto::Event::PeerCreated
-    /// [`Event::TracksApplied`]: medea_client_api_proto::Event::TracksApplied
+    /// [`Event::PeerUpdated`]: medea_client_api_proto::Event::PeerUpdated
     fn handle(
         &mut self,
         msg: NegotiationNeeded,